Began exploratory data analysis to explore relationship between revenue and other variables.
Currently, I am trying to plot differences in the distributions for all continuous variables in the dataset under the two conditions - no revenue or revenue. To start this preliminary analysis, I tried to make interactive plots using Plotly, as there is large dynamic range between different variables, and it would be useful to explore the variables in more depth. I managed to make the plot for the revenue labels. However, the code does not show the no revenue variable. That is something I will have to look into.